+++ NEWS.md
@@ -3,6 +3,7 @@
 - Breaking: Heavy remanufacturing of `define_variance_wrapper`
 
     * New: `technical_data` argument offers a more consistent way to include technical data within the enclosing environment of the wrapper. `objects_to_include` is kept for non-data objects (such as additional linearization wrappers) or advanced customization.
+    * New: `technical_param` argument offers a more convenient way to specify default values for parameters used by the variance function.
     * New: `reference_weight` replaces `default$weight`. This means that the reference weight used for point estimation and linearization is set while defining the variance wrapper and not at run-time.
     * Change: `stat`, which was a remain of an early implementation of linearization functions, is not a parameter of the variance wrappers anymore. Its purpose (to apply a given variance wrapper to several variables without having to type the name of the linearization wrapper) should be covered in a more consistent way by a forthcoming escape hatch allowing standard evaluation.
     * Deprecated: `default` is replaced by `default_id`, as `default$weight` and `default$stat` are no longer needed. As for `default$alpha`, its value is set to 0.05 and cannot be changed anymore while defining the variance wrapper (as this can easily be done afterwards using `formals<-`).